+++ user-management/src/main/java/org/shaft/administration/usermanagement/repositories/fingerprint/IdentityCustomRepository.java
@@ -12,6 +12,6 @@ public interface IdentityCustomRepository {
     Flux<Identity> checkIfFpExistsForI(String fp, int i, boolean isIdentified);
     Flux<Identity> checkIfIExistsForFp(String fp);
     Mono<Long> updateFp(String fp, int i);
-    Mono<Long>  upsertFpAndIPair(int account, String fp, int i);
+    Mono<Long>  upsertFpAndIPair(int account, String fp, long i);
     Mono<Identity> save(int accountId,Identity i);
 }
